order_manager/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
groq==0.4.1
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
pillow==10.1.0
